Adds loading state and delete functionality to users/orders in admin dashboard

+++ app/(auth)/(components)/Action.tsx
@@ -5,25 +5,41 @@ import { useRouter } from 'next/navigation';
 import axios from "axios";
 import { message } from 'antd';
 
-const Action = ({ value, onDelete }: { value: string, onDelete: (orderId: string) => void }) => {
+const Action = ({ value, onDelete, type }: { value: string, onDelete: (orderId: string) => void, type:string }) => {
     const router = useRouter();
 
     const deleteHandler = async () => {
-    const confirmDelete = window.confirm("Are you sure you want to delete this order?");
+    const confirmDelete = window.confirm("Are you sure you want to delete this?");
     if (!confirmDelete) return;
+    if(type === "orders") {
+        try {
+            await axios.delete(`/api/delete-order`, {
+                params: { orderId: value }, 
+            });
+            onDelete(value);
+            message.success("Order deleted successfully");
+            router.refresh();
 
-    try {
-        await axios.delete(`/api/delete-order`, {
-            params: { orderId: value }, 
-        });
-        onDelete(value);
-        message.success("Order deleted successfully");
-        router.refresh();
-
-    } catch (error) {
-        console.error("Error deleting order:", error);
-        message.error("Something went wrong: " + error);
+        } catch (error) {
+            console.error("Error deleting order:", error);
+            message.error("Something went wrong: " + error);
+        } 
     }
+    if(type === "users"){
+        try {
+            await axios.delete(`/api/delete-user`, {
+                params: { userId: value }, 
+            });
+            onDelete(value);
+            message.success("User deleted successfully");
+            router.refresh();
+
+        } catch (error) {
+            console.error("Error deleting user:", error);
+            message.error("Something went wrong: " + error);
+        }
+    }
+    
 };
 
 
@@ -34,7 +50,7 @@ const Action = ({ value, onDelete }: { value: string, onDelete: (orderId: string
             </DropdownMenuTrigger>
             <DropdownMenuContent>
                 <DropdownMenuItem>
-                    <div className="flex flex-row text-lg gap-2 cursor-pointer" onClick={() => router.push(`/dashboard/category/${value}`)}>
+                    <div className="flex flex-row text-lg gap-2 cursor-pointer" onClick={() => router.push(`/dashboard/${type}/${value}`)}>
                         <Edit />
                         <p>Edit</p>
                     </div>
